Added absoluteweather.csv Will explore plotting mV instead of normalized - I think we should use mV, it's easier to identify differences in reactions and compare to other studies. Apparently the SYNTECH machine should already normalize your responses... so that would have been nice to know.
Phill suggested edits for the filtering of odours that we didn't like (using tidyverse)

We should plot a simple graph of humidity versus mV to see if there's any correlation, but I guess we could also plot mV versus time or date - I need to add that data into the file if that is something to compare. Natually I'd expect the specimine to dry over time, so responses get noisier (possibly larger mV recordings) and I think this is part of what the SYNTECH program "normalization" accounts for (not sure if it's working).